The bash script now has a debug option

+++ run.sh
@@ -9,4 +9,7 @@ fi
 cat $1 | sort > /tmp/offisort;
 cat $1 | bin/sort > /tmp/mysort;
 diff /tmp/offisort /tmp/mysort;
-rm /tmp/offisort /tmp/mysort;
+if [ $# -lt 2 ]
+then
+	rm /tmp/offisort /tmp/mysort;
+fi
